write better code
